move addViewNotifyConfig() to NotificationVS
