+++ gsn-tiny/app/src/main/java/tinygsn/gui/android/ActivityVSConfig.java
@@ -25,31 +25,14 @@
 
 package tinygsn.gui.android;
 
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Properties;
-
-import tinygsn.beans.DataField;
-import tinygsn.beans.StaticData;
-import tinygsn.beans.StreamSource;
-import tinygsn.controller.AndroidControllerVS;
-import tinygsn.model.vsensor.AbstractVirtualSensor;
-import tinygsn.model.vsensor.NotificationVirtualSensor;
-import tinygsn.model.wrappers.AbstractWrapper;
-import tinygsn.storage.db.SqliteStorageManager;
-
 import android.app.Activity;
-import android.app.AlertDialog;
 import android.content.Context;
 import android.graphics.Color;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.text.Editable;
 import android.text.InputType;
-import android.text.TextWatcher;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemSelectedListener;
@@ -64,8 +47,20 @@ import android.widget.TableRow;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import android.view.Menu;
-import android.view.MenuItem;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Properties;
+
+import tinygsn.beans.DataField;
+import tinygsn.beans.StaticData;
+import tinygsn.beans.StreamSource;
+import tinygsn.controller.AndroidControllerVS;
+import tinygsn.model.vsensor.AbstractVirtualSensor;
+import tinygsn.model.wrappers.AbstractWrapper;
+import tinygsn.storage.db.SqliteStorageManager;
 
 public class ActivityVSConfig extends Activity {
 	static int TEXT_SIZE = 10;
@@ -150,8 +145,12 @@ public class ActivityVSConfig extends Activity {
 			public void onItemSelected(AdapterView<?> parent, View view, final int pos, long id) {
 				table_vsensor_config.removeAllViews();
 
-				if (pos == 1) addViewNotifyConfig();
-				else table_notify_config.removeAllViews();
+				try {
+					table_notify_config.removeAllViews();
+					((AbstractVirtualSensor) Class.forName(AbstractVirtualSensor.VIRTUAL_SENSOR_CLASSES[pos]).newInstance()).getRowParameters(table_notify_config, context);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 
 				new AsyncTask<Activity, Void, SettingPanel>() {
 					@Override
@@ -350,194 +349,12 @@ public class ActivityVSConfig extends Activity {
 		return row;
 	}
 
-	public void addViewNotifyConfig() { //TODO remove and move parameters to VS parameters
-		table_notify_config.removeAllViews();
-
-		TableRow row = new TableRow(this);
-
-		// Row Field
-		TextView txt = new TextView(this);
-		txt.setText("Field");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		field = new Spinner(this);
-		List<String> list = new ArrayList<String>();
-		String wrapperName = wrapperList.getProperty("gps");
-
-		try {
-			AbstractWrapper w = (AbstractWrapper) StaticData.getWrapperByName(wrapperName);
-			for (String s : w.getFieldList()) {
-				list.add(s);
-			}
-		} catch (Exception e1) {
-			e1.printStackTrace();
-		}
-
-		ArrayAdapter<String> dataAdapter = new ArrayAdapter<String>(this,
-				                                                           R.layout.spinner_item, list);
-		dataAdapter
-				.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-		field.setAdapter(dataAdapter);
-
-		row.addView(field);
-		table_notify_config.addView(row);
-
-		// Row condition
-		row = new TableRow(this);
-
-		txt = new TextView(this);
-		txt.setText("Sampling Rate");
-		txt.setText("Condition    ");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		condition = new Spinner(this);
-		List<String> list_condition = new ArrayList<String>();
-
-		for (String s : NotificationVirtualSensor.CONDITIONS) {
-			list_condition.add(s);
-		}
-		ArrayAdapter<String> dataAdapter_condition = new ArrayAdapter<String>(this,
-				                                                                     R.layout.spinner_item, list_condition);
-		dataAdapter
-				.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-		condition.setAdapter(dataAdapter_condition);
-		row.addView(condition);
-
-		table_notify_config.addView(row);
-
-		// Row value
-		row = new TableRow(this);
-
-		txt = new TextView(this);
-		txt.setText("Value");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		editText_value = new EditText(this);
-		editText_value.setText("10");
-		editText_value.setTextSize(TEXT_SIZE + 5);
-		editText_value.setInputType(InputType.TYPE_CLASS_NUMBER);
-		// editText_value.requestFocus();
-		editText_value.setTextColor(Color.parseColor("#000000"));
-		row.addView(editText_value);
-
-		editText_value.addTextChangedListener(new TextWatcher() {
-			@Override
-			public void onTextChanged(CharSequence s, int start, int before, int count) {
-			}
-
-			@Override
-			public void beforeTextChanged(CharSequence s, int start, int count,
-			                              int after) {
-			}
-
-			@Override
-			public void afterTextChanged(Editable s) {
-				try {
-					// numLatest =
-					// Integer.parseInt(editText_numLatest.getText().toString());
-					// loadLatestData();
-				} catch (NumberFormatException e) {
-					AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(
-							                                                                context);
-					alertDialogBuilder.setTitle("Please input a number!");
-				}
-			}
-		});
-
-		table_notify_config.addView(row);
-
-		// Row action
-		row = new TableRow(this);
-
-		txt = new TextView(this);
-		txt.setText("Action");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		action = new Spinner(this);
-		List<String> list_action = new ArrayList<String>();
-
-		for (String s : NotificationVirtualSensor.ACTIONS) {
-			list_action.add(s);
-		}
-		ArrayAdapter<String> dataAdapter_action = new ArrayAdapter<String>(this,
-				                                                                  R.layout.spinner_item, list_action);
-		dataAdapter
-				.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-		action.setAdapter(dataAdapter_action);
-		row.addView(action);
-
-		table_notify_config.addView(row);
-
-		// Row contact
-		row = new TableRow(this);
-
-		txt = new TextView(this);
-		txt.setText("Contact");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		editText_contact = new EditText(this);
-		editText_contact.setText("+41798765432");
-		editText_contact.setTextSize(TEXT_SIZE + 5);
-		// editText_contact.setInputType(InputType.TYPE_CLASS_NUMBER);
-		// editText_contact.requestFocus();
-		editText_contact.setTextColor(Color.parseColor("#000000"));
-		row.addView(editText_contact);
-
-		editText_contact.addTextChangedListener(new TextWatcher() {
-			@Override
-			public void onTextChanged(CharSequence s, int start, int before, int count) {
-			}
-
-			@Override
-			public void beforeTextChanged(CharSequence s, int start, int count,
-			                              int after) {
-			}
-
-			@Override
-			public void afterTextChanged(Editable s) {
-				try {
-					// numLatest =
-					// Integer.parseInt(editText_numLatest.getText().toString());
-					// loadLatestData();
-				} catch (NumberFormatException e) {
-					AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(
-							                                                                context);
-					alertDialogBuilder.setTitle("Please input a phone number!");
-				}
-			}
-		});
-
-		table_notify_config.addView(row);
-
-		// Row Save to DB
-		row = new TableRow(this);
-
-		txt = new TextView(this);
-		txt.setText("Save to Database?");
-		txt.setTextColor(Color.parseColor("#000000"));
-		row.addView(txt);
-
-		saveToDB = new CheckBox(this);
-		saveToDB.setTextColor(Color.parseColor("#000000"));
-		row.addView(saveToDB);
-
-
-		table_notify_config.addView(row);
-		// TableRow.LayoutParams params = new TableRow.LayoutParams();
-		// params.span = 2;
-
-	}
-
 	@Override
 	protected void onPause() {
 		super.onPause();
 	}
 
+	//FIXME : think about how to create the row for db
 	public void saveVS() {
 		String vsName = editText_vsName.getText().toString();
 		String notify_field = "", notify_condition = "", notify_value = "", notify_action = "", notify_contact = "", save_to_db = "";
